@@ -1,10 +1,10 @@
 name: tiledjsonreader
 description: Reader of the Tiled file .json and your tileSet.json
-version: 0.0.8
+version: 1.0.0
 homepage: https://github.com/RafaelBarbosatec/tiled_json_reader
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
