@@ -6,52 +6,53 @@ import 'package:tiledjsonreader/map/tile_set_detail.dart';
 import 'package:tiledjsonreader/util/extensions.dart';
 
 class TiledMap {
-  int compressionLevel;
-  int height;
-  bool infinite;
-  List<MapLayer> layers;
-  int nextLayerId;
-  int nextObjectId;
-  String orientation;
-  String renderOrder;
-  String tiledVersion;
-  int tileHeight;
-  List<TileSetDetail> tileSets;
-  int tileWidth;
-  String type;
-  double version;
-  int width;
+  int? compressionLevel;
+  int? height;
+  bool? infinite;
+  List<MapLayer>? layers;
+  int? nextLayerId;
+  int? nextObjectId;
+  String? orientation;
+  String? renderOrder;
+  String? tiledVersion;
+  int? tileHeight;
+  List<TileSetDetail>? tileSets;
+  int? tileWidth;
+  String? type;
+  double? version;
+  int? width;
 
-  TiledMap(
-      {this.compressionLevel,
-      this.height,
-      this.infinite,
-      this.layers,
-      this.nextLayerId,
-      this.nextObjectId,
-      this.orientation,
-      this.renderOrder,
-      this.tiledVersion,
-      this.tileHeight,
-      this.tileSets,
-      this.tileWidth,
-      this.type,
-      this.version,
-      this.width});
+  TiledMap({
+    this.compressionLevel,
+    this.height,
+    this.infinite,
+    this.layers,
+    this.nextLayerId,
+    this.nextObjectId,
+    this.orientation,
+    this.renderOrder,
+    this.tiledVersion,
+    this.tileHeight,
+    this.tileSets,
+    this.tileWidth,
+    this.type,
+    this.version,
+    this.width,
+  });
 
   TiledMap.fromJson(Map<String, dynamic> json) {
     compressionLevel = json['compressionlevel'];
     height = json['height'];
     infinite = json['infinite'];
     if (json['layers'] != null) {
-      layers = new List<MapLayer>();
+      layers = <MapLayer>[];
       json['layers'].forEach((v) {
         if (v['type'] == TypeLayer.tilelayer.getName()) {
-          layers.add(TileLayer.fromJson(v));
+          layers?.add(TileLayer.fromJson(v));
         } else if (v['type'] == TypeLayer.objectgroup.getName()) {
-          layers.add(ObjectGroup.fromJson(v));
+          layers?.add(ObjectGroup.fromJson(v));
         } else {
-          layers.add(MapLayer.fromJson(v));
+          layers?.add(MapLayer.fromJson(v));
         }
       });
     }
@@ -62,9 +63,9 @@ class TiledMap {
     tiledVersion = json['tiledversion'];
     tileHeight = json['tileheight'];
     if (json['tilesets'] != null) {
-      tileSets = new List<TileSetDetail>();
+      tileSets = <TileSetDetail>[];
       json['tilesets'].forEach((v) {
-        tileSets.add(new TileSetDetail.fromJson(v));
+        tileSets?.add(new TileSetDetail.fromJson(v));
       });
     }
     tileWidth = json['tilewidth'];
@@ -79,7 +80,7 @@ class TiledMap {
     data['height'] = this.height;
     data['infinite'] = this.infinite;
     if (this.layers != null) {
-      data['layers'] = this.layers.map((v) => v.toJson()).toList();
+      data['layers'] = this.layers?.map((v) => v.toJson()).toList();
     }
     data['nextlayerid'] = this.nextLayerId;
     data['nextobjectid'] = this.nextObjectId;
@@ -88,7 +89,7 @@ class TiledMap {
     data['tiledversion'] = this.tiledVersion;
     data['tileheight'] = this.tileHeight;
     if (this.tileSets != null) {
-      data['tilesets'] = this.tileSets.map((v) => v.toJson()).toList();
+      data['tilesets'] = this.tileSets?.map((v) => v.toJson()).toList();
     }
     data['tilewidth'] = this.tileWidth;
     data['type'] = this.type;
