@@ -1,35 +1,36 @@
 import 'package:tiledjsonreader/tile_set/tile_set_object.dart';
 
 class TileSetObjectGroup {
-  String drawOrder;
-  int id;
-  String name;
-  List<TileSetObject> objects;
-  int opacity;
-  String type;
-  bool visible;
-  int x;
-  int y;
+  String? drawOrder;
+  int? id;
+  String? name;
+  List<TileSetObject>? objects;
+  int? opacity;
+  String? type;
+  bool? visible;
+  int? x;
+  int? y;
 
-  TileSetObjectGroup(
-      {this.drawOrder,
-      this.id,
-      this.name,
-      this.objects,
-      this.opacity,
-      this.type,
-      this.visible,
-      this.x,
-      this.y});
+  TileSetObjectGroup({
+    this.drawOrder,
+    this.id,
+    this.name,
+    this.objects,
+    this.opacity,
+    this.type,
+    this.visible,
+    this.x,
+    this.y,
+  });
 
   TileSetObjectGroup.fromJson(Map<String, dynamic> json) {
     drawOrder = json['draworder'];
     id = json['id'];
     name = json['name'];
     if (json['objects'] != null) {
-      objects = new List<TileSetObject>();
+      objects = <TileSetObject>[];
       json['objects'].forEach((v) {
-        objects.add(new TileSetObject.fromJson(v));
+        objects?.add(new TileSetObject.fromJson(v));
       });
     }
     opacity = json['opacity'];
@@ -45,7 +46,7 @@ class TileSetObjectGroup {
     data['id'] = this.id;
     data['name'] = this.name;
     if (this.objects != null) {
-      data['objects'] = this.objects.map((v) => v.toJson()).toList();
+      data['objects'] = this.objects?.map((v) => v.toJson()).toList();
     }
     data['opacity'] = this.opacity;
     data['type'] = this.type;
