@@ -34,7 +34,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: Text(widget.title ?? ''),
       ),
       body: Container(),
       floatingActionButton: FloatingActionButton(
@@ -48,7 +48,7 @@ class _MyHomePageState extends State<MyHomePage> {
   void _loadMap() {
     TiledJsonReader tiled = TiledJsonReader('assets/mapa1.json');
     tiled.read().then((value) {
-      print(value);
+      print(value.toJson());
     });
   }
 }
